refactor(redskilled): what a reader asked for is not what the answer contains
The remote-counter block pushed `statusline-payload.ts` past its shrink-only
baseline, and the withholding is the part that was never about the document:
`statusline-payload` decides what one answer CONTAINS, and this decides how much
of that answer one reader asked for.

It moves to `statusline-extras.ts` whole — the extras vocabulary, the opt-in
request, the withholding and the two honest-absence constants — and is
re-exported from the payload module, because the payload interface names the
extras in `withheld` and every caller of the withholding already imports the
payload it withholds from. No importer changes.

Refs #3562
