+++ apps/redskilled/src/statusline-payload.ts
@@ -32,16 +32,17 @@ import type { RedskilledHostState, RedskilledRssSource, RedskilledWorkerView } f
 import { isRedskilledStatuslineMetrics, type RedskilledStatuslineMetrics } from "./live-metrics.js";
 import { resolveEnforcedBudget, type RedskilledBudgetName, type RedskilledRssReading } from "./memory-sampler.js";
 import {
+  buildActivityReport,
+  isRedskilledActivityReport,
+  type RedskilledActivityReport,
+} from "./activity-report.js";
+import {
   buildRemoteCounterReport,
   isRedskilledRemoteCounterReport,
   type RedskilledRemoteCounterReport,
 } from "./remote-counters.js";
-import {
-  buildActivityReport,
-  isRedskilledActivityReport,
-  type RedskilledActivityReport,
-  type RedskilledRepositoryActivity,
-} from "./repository-activity.js";
+import type { RedskilledRepositoryActivity } from "./repository-activity.js";
+import type { RedskilledStatuslineExtra } from "./statusline-extras.js";
 import type { RedskilledWorkerDisplay, RedskilledWorkerDisplayRecord } from "./worker-display.js";
 import {
   buildDeaths,
@@ -75,6 +76,17 @@ export {
   type RedskilledStatuslineDeath,
   type RedskilledStatuslineDeaths,
 } from "./statusline-deaths.js";
+
+// How much of this document one reader asked for is its own question, answered
+// in `./statusline-extras.js`. Re-exported because the payload interface NAMES
+// the extras — `withheld` is typed by them — and because every caller of the
+// withholding already imports this module for the payload it withholds from.
+export {
+  REDSKILLED_STATUSLINE_EXTRAS,
+  withholdStatuslineExtras,
+  type RedskilledStatuslineExtra,
+  type RedskilledStatuslineExtrasRequest,
+} from "./statusline-extras.js";
 
 // The remote-counter block's own judgements — which counters exist, when an age
 // makes one stale, what an absence reads as — live in `./remote-counters.js`.
@@ -405,78 +417,6 @@ export interface RedskilledStatuslinePayload {
    */
   readonly withheld?: readonly RedskilledStatuslineExtra[];
 }
-
-/**
- * One block that scales with Worker count, and so travels on request.
- *
- * Named individually rather than as one `verbose` boolean because the surfaces
- * want different subsets: a statusline wants vitals and no logs, a dashboard
- * wants the display records, and a health probe wants none of the three.
- */
-export type RedskilledStatuslineExtra = "logs" | "vitals" | "display";
-
-/** Every extra there is, so a caller can ask for the skeleton by subtracting. */
-export const REDSKILLED_STATUSLINE_EXTRAS: readonly RedskilledStatuslineExtra[] = ["logs", "vitals", "display"];
-
-/**
- * Which extras a reader wants; an omitted flag is a block it does not need.
- *
- * A record of opt-INS rather than opt-outs: the expensive direction should be
- * the one a caller had to type.
- */
-export interface RedskilledStatuslineExtrasRequest {
-  readonly logs?: boolean;
-  readonly vitals?: boolean;
-  readonly display?: boolean;
-}
-
-/**
- * The same payload with the extras nobody asked for removed. PURE.
- *
- * **A withheld block is replaced by its own honest absence, never deleted**: the
- * shape stays total, so a consumer written against the full document renders a
- * skeleton response without a single existence check. What it must not do is
- * read the absence as a measurement — which is exactly what `withheld` is for.
- *
- * `undefined` extras means the whole document, because that is what every client
- * pinned to an older bundle asks for by saying nothing (ADR 0130 rule 3). A
- * caller that wants less says so.
- */
-export function withholdStatuslineExtras(
-  payload: RedskilledStatuslinePayload,
-  extras: RedskilledStatuslineExtrasRequest | undefined,
-): RedskilledStatuslinePayload {
-  if (extras === undefined) return payload;
-  const withheld = REDSKILLED_STATUSLINE_EXTRAS.filter((extra) => extras[extra] !== true);
-  if (withheld.length === 0) return payload;
-  const keep = (extra: RedskilledStatuslineExtra) => extras[extra] === true;
-  return {
-    ...payload,
-    workers: payload.workers.map((worker) => ({
-      ...worker,
-      ...(keep("vitals") ? {} : { vitals: WITHHELD_VITALS, budget: { ...worker.budget, used_bytes: null, used_fraction: null } }),
-      ...(keep("logs") ? {} : { log: WITHHELD_LOG }),
-      ...(keep("display") ? {} : { display: null, display_published_at: null }),
-    })),
-    withheld,
-  };
-}
-
-/** The vitals of a Worker nobody asked about — total in shape, empty in fact. */
-const WITHHELD_VITALS: RedskilledStatuslineVitals = {
-  rss_bytes: null,
-  sampled_at: null,
-  age_ms: null,
-  fresh: false,
-  rss_source: null,
-};
-
-/** The log of a Worker nobody asked about. `null`, exactly as an unpublished one. */
-const WITHHELD_LOG: RedskilledStatuslineWorkerLog = {
-  last_line: null,
-  published_at: null,
-  source: null,
-};
 
 export interface BuildStatuslinePayloadInput {
   readonly hostState: RedskilledHostState;
